Разделяет проект на модули
