+++ js/consts.js
@@ -1,0 +1,8 @@
+export default {
+  PHOTO_COUNT: 25,
+  COMMENT_MAX_COUNT: 10,
+  AVATAR_MIN_VALUE: 1,
+  AVATAR_MAX_VALUE: 6,
+  LIKES_MIN_VALUE: 15,
+  LIKES_MAX_VALUE: 255
+};
