+++ js/photo/show-photo.js
@@ -1,0 +1,1 @@
+// модуль для интерфейса просмотра фото: модальное окно
