+++ js/photo/upload-photo.js
@@ -1,0 +1,1 @@
+// модуль для интерфейса загрузки изображения
