+++ js/photo/zoom-photo.js
@@ -1,0 +1,1 @@
+// модуль для интерфейса мамштабирования изображения
