+++ js/photoDescription/add-comment.js
@@ -1,0 +1,1 @@
+// модуль для интерфейса добавления комментария
