+++ js/photoDescription/add-hashtag.js
@@ -1,0 +1,1 @@
+// модуль для интерфейса добавления хэштегов
