+++ js/photoList/filter-list.js
@@ -1,0 +1,1 @@
+// модуль для интерфейса фильтрации списка
