+++ js/photoList/photo-list.js
@@ -1,0 +1,1 @@
+// модуль для интерфейса показа списка фото
